+++ ecdsa_test.go
@@ -30,22 +30,6 @@ var testKeys = [...]string{
 	"02a33413277a319cc6fd4c54a2feb9032eba668ec587f307e319dc48733087fa61",
 }
 
-func Test_decompressPoints(t *testing.T) {
-	t.Run("prepared public keys: decompressPoints", func(t *testing.T) {
-		for i := range testKeys {
-			bytes, err := hex.DecodeString(testKeys[i])
-			require.NoErrorf(t, err, testKeys[i])
-
-			x, y := decompressPoints(new(big.Int).SetBytes(bytes[1:]), uint(bytes[0]))
-			require.NotNil(t, x)
-			require.NotNil(t, y)
-
-			res := MarshalPublicKey(&ecdsa.PublicKey{Curve: curve, X: x, Y: y})
-			require.Equal(t, testKeys[i], hex.EncodeToString(res))
-		}
-	})
-}
-
 func TestMarshalUnmarshal(t *testing.T) {
 	t.Run("prepared public keys: unmarshal / marshal", func(t *testing.T) {
 		for i := range testKeys {
