+++ go.mod
@@ -3,8 +3,8 @@ module github.com/nspcc-dev/neofs-crypto
 go 1.13
 
 require (
-	github.com/mr-tron/base58 v1.1.2
+	github.com/mr-tron/base58 v1.2.0
 	github.com/nspcc-dev/rfc6979 v0.2.0
-	github.com/pkg/errors v0.8.1
-	github.com/stretchr/testify v1.4.0
+	github.com/pkg/errors v0.9.1
+	github.com/stretchr/testify v1.7.0
 )
